+++ src/importers/federated_learning.cpp
@@ -55,7 +55,7 @@ std::vector<double> aggregateElementWiseMedian(
         std::sort(values.begin(), values.end());
         const std::size_t mid = values.size() / 2;
         out[d] = values.size() % 2 == 0
-            ? (values[mid - 1] + values[mid]) / 2.0
+            ? (values[static_cast<int>(mid - 1)] + values[mid]) / 2.0
             : values[mid];
     }
     return out;
@@ -150,7 +150,7 @@ json FederatedImportCoordinator::FederatedAggregator::aggregateUpdates(const std
             if (aggregation_algorithm == "median" && !values.empty()) {
                 std::sort(values.begin(), values.end());
                 size_t mid      = values.size() / 2;
-                aggregated[key] = values.size() % 2 == 0 ? (values[mid - 1] + values[mid]) / 2.0 : values[mid];
+                aggregated[key] = values.size() % 2 == 0 ? (values[static_cast<int>(mid - 1)] + values[mid]) / 2.0 : values[mid];
             } else if (aggregation_algorithm == "trimmed_mean" && values.size() >= 3) {
                 std::sort(values.begin(), values.end());
                 // trim one min and one max when possible (Byzantine-robust default)
